Refactor Firebase imports to use consistent casing and update styles import paths. Remove deprecated ApplicationsTable component and ensure all screens utilize the new Firebase configuration.
