+++ src/Screens/ApplicationForm.tsx
@@ -1,14 +1,14 @@
 import React, { useState, useEffect } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 import { doc, getDoc, addDoc, updateDoc, collection } from 'firebase/firestore';
-import { auth, db } from '../Config/firebase';
+import { auth, db } from '../Firebase/firebase';
 import { JobApplication } from '../Types';
 import { JOB_TYPES, APPLICATION_STATUS } from '../Constants';
 import Button from '../Components/Button';
 import Select from '../Components/Select';
 import CustomTextField from '../Components/CustomTextField';
 import { faPlus, faTimes, faSave, faTimes as faCancel } from '@fortawesome/free-solid-svg-icons';
-import styles from '../styles/form.module.css';
+import styles from '../Styles/form.module.css';
 
 const ApplicationForm: React.FC = () => {
   const navigate = useNavigate();
