+++ src/Screens/Dashboard.tsx
@@ -1,12 +1,12 @@
 import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { collection, query, where, onSnapshot } from 'firebase/firestore';
-import { auth, db } from '../Config/firebase';
+import { auth, db } from '../Firebase/firebase';
 import { JobApplication } from '../Types';
 import Header from '../Components/Header';
 import ApplicationTable from '../Components/ApplicationTable';
 import KPICards from '../Components/KPICards';
-import styles from '../styles/dashboard.module.css';
+import styles from '../Styles/dashboard.module.css';
 
 const Dashboard: React.FC = () => {
   const [applications, setApplications] = useState<JobApplication[]>([]);
