+++ src/Screens/Login.tsx
@@ -2,11 +2,11 @@ import React, { useState } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
 import { signInWithEmailAndPassword } from 'firebase/auth';
 import { doc, updateDoc } from 'firebase/firestore';
-import { auth, db } from '../Config/firebase';
+import { auth, db } from '../Firebase/firebase';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faEye, faEyeSlash } from '@fortawesome/free-solid-svg-icons';
-import styles from '../styles/auth.module.css';
-import componentStyles from '../styles/components.module.css';
+import styles from '../Styles/auth.module.css';
+import componentStyles from '../Styles/components.module.css';
 import { motion } from 'framer-motion';
 import CustomButton from '../Components/CustomButton';
 import Loader from '../Components/Loader';
