+++ src/components/ApplicationTable.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useMemo } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { deleteDoc, doc } from 'firebase/firestore';
-import { db } from '../Config/firebase';
+import { db } from '../Firebase/firebase';
 import { JobApplication } from '../Types';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {
@@ -16,7 +16,7 @@ import {
   faCheck,
   faTimes
 } from '@fortawesome/free-solid-svg-icons';
-import styles from '../styles/applicationTable.module.css';
+import styles from '../Styles/applicationTable.module.css';
 
 interface ApplicationTableProps {
   applications: JobApplication[];
