+++ src/components/Header.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { doc, getDoc } from 'firebase/firestore';
-import { auth, db } from '../Config/firebase';
+import { auth, db } from '../Firebase/firebase';
 import { UserProfile } from '../Types';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {
@@ -10,7 +10,7 @@ import {
   faChartLine,
 } from '@fortawesome/free-solid-svg-icons';
 import ProfileMenu from './ProfileMenu';
-import styles from '../styles/header.module.css';
+import styles from '../Styles/header.module.css';
 
 const Header: React.FC = () => {
   const navigate = useNavigate();
