+++ src/components/ProfileMenu.tsx
@@ -1,11 +1,11 @@
 import React, { useState, useRef, useEffect } from 'react';
 import { signOut } from 'firebase/auth';
 import { useNavigate } from 'react-router-dom';
-import { auth } from '../Config/firebase';
+import { auth } from '../Firebase/firebase';
 import { UserProfile } from '../Types';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faSignOut } from '@fortawesome/free-solid-svg-icons';
-import styles from '../styles/profileMenu.module.css';
+import styles from '../Styles/profileMenu.module.css';
 
 interface ProfileMenuProps {
   user: UserProfile;
