+++ src/components/SplashScreen.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useAuth } from '../Contexts/AuthContext';
-import styles from '../styles/splashScreen.module.css';
+import styles from '../Styles/splashScreen.module.css';
 
 const SplashScreen: React.FC = () => {
   const navigate = useNavigate();
